+++ CHANGELOG.md
@@ -11,7 +11,7 @@
   `PossiblyCurrentContext::make_current_surfaceless(&self)` in the `Wgl`
   implementation to allow the use of surfaceless contexts with WGL.
 - Added workaround for EGL drivers reporting `EGL_KHR_platform_gbm` without EGL 1.5 client.
-
+- Added `GlContext::priority`/`ContextAttributesBuilder::with_priority` to get/set context priority.
 
 # Version 0.32.1